pyproject.toml
[project]
name = "yatackerhelper"
version = "0.6.0"
description = "Telegram бот для полного клонирования проектов из Yandex Tracker со всеми связанными сущностями"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.22.0",
    "aiogram-dialog>=2.4.0",
    "alembic>=1.16.5",
    "apscheduler>=3.11.2",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.43",
    "yatrackerapi==2.1.3",
    "python-fasthtml>=0.6.0",
]

[project.urls]
Homepage = "https://github.com/imdeniil/YaTackerHelper"
Repository = "https://github.com/imdeniil/YaTackerHelper"
Issues = "https://github.com/imdeniil/YaTackerHelper/issues"

[project.authors]
name = "Daniil Pavliuchik"
email = "keemor821@gmail.com"
